+++ CHANGELOG.md
@@ -8,6 +8,7 @@ NOTES:
 * Added protection group resource
 * Added Importers to all resources
 * Added ability to update all resource parameters
+* Added some random number to names in tests, so they can be run multiple times without cleaning up
 
 
 ## 0.1.0 (December 2, 2018)